Initial commit of public version
The current version has certain information hardcoded, such as names,
office addresses, phone numbers, etc.  Out of an abundance of caution,
that information has been edited out so that it is not readily available
on a public repository.  Such information will eventually be placed in
some sort of settings file or local DB.